services/github.py
import base64
import datetime
import json
import requests
import yaml
import os
from github import Github, Auth

# using an access token
auth = Auth.Token(os.getenv("ORG_TOKEN"))
# First create a Github instance:
# Public Web Github
gh = Github(auth=auth)
