requirements.txt
click==8.1.7
cryptography==41.0.4
PyGithub==2.1.1
pyjwt==2.8.0
PyYAML==6.0.1